Rename internal scope mod `scoperef` -> `sref`.
